Move homepage to root address and login in directory - and change all related redirects
