+++ app/page.tsx
@@ -1,68 +1,51 @@
 "use client"
+import { Container, Box, Heading, Text } from "@chakra-ui/react";
+import { MenuGrid } from "@/app/login/components/menu-grid";
+import { useRequireAuth } from "@/lib/auth-context"
+import {AuthWrapper} from "@/components/auth-wrapper";
 
-import React, { useState } from "react"
-import { useAuth } from "@/lib/auth-context"
-import { Input, Button, VStack, Text, Container } from "@chakra-ui/react"
-import { useRouter } from "next/navigation"
+function HomepageContent() {
+  const { player, loading } = useRequireAuth()
 
-export default function LoginPage() {
-  const { login, user } = useAuth()
-  const router = useRouter()
-  const [username, setUsername] = useState("")
-  const [password, setPassword] = useState("")
-  const [loading, setLoading] = useState(false)
-
-  // Redirect if already logged in
-  React.useEffect(() => {
-    if (user) {
-      router.push("/dashboard")
-    }
-  }, [user, router])
-
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault()
-    setLoading(true)
-
-    try {
-      await login(username, password)
-      router.push("/dashboard")
-    } catch (error: any) {
-      console.log(error.message || "Login failed")
-    } finally {
-      setLoading(false)
-    }
+  if (loading) {
+    return (
+      <Box minH="100vh" bg="black" p={4} display="flex" alignItems="center" justifyContent="center">
+        <Text color="white">Loading...</Text>
+      </Box>
+    )
   }
 
-  return (
-    <Container maxWidth="4xl" centerContent py={20}>
-      <VStack gap={6} align="center" maxW="sm" w="full">
-        <Text fontSize="4xl" fontWeight="light" color="white">
-          Footage Tracker
-        </Text>
+  if (!player) {
+    return null
+  }
 
-        <form onSubmit={handleLogin} style={{ width: "100%" }}>
-          <VStack gap={4} w="full">
-            <Input
-              placeholder="Username"
-              value={username}
-              onChange={(e) => setUsername(e.target.value)}
-              autoComplete="username"
-              required
-            />
-            <Input
-              placeholder="Password"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              autoComplete="current-password"
-              required
-            />
-            <Button type="submit" colorScheme="green" loading={loading} disabled={loading} w="full">
-              Login
-            </Button>
-          </VStack>
-        </form>
-      </VStack>
-    </Container>
+  const menuItems = [
+    { title: "Sources", icon: "database", href: "/sources" },
+    { title: "Events", icon: "calendar", href: "/events" },
+    { title: "Clips", icon: "film", href: "/clips" },
+    { title: "Points", icon: "scoreboard", href: "/points" },
+    { title: "Stats", icon: "bar-chart", href: "/stats" },
+    { title: "Playlists", icon: "playlist", href: "/playlists" },
+    ...(player.is_admin ? [{ title: "Players", icon: "users", href: "/players" }] : []),
+  ];
+
+  return (
+    <Box minH="100vh" bg="black" p={4}>
+      <Container maxW="4xl">
+        <Heading as="h1" size="4xl" fontWeight="light" color="white" mb={8} mt={4}>
+          Hello, {player.username}
+        </Heading>
+        <MenuGrid items={menuItems} />
+      </Container>
+    </Box>
+  );
+}
+
+export default function Homepage() {
+  return (
+    <AuthWrapper>
+      <HomepageContent />
+    </AuthWrapper>
   )
 }
+
